requirements.txt
Django>=4.0
djangorestframework>=3.13.1
django-celery-beat>=2.2.1
celery>=5.2.3
django-environ>=0.8
gunicorn>=20.1.0
django-allauth
pillow~=10.4.0
locust~=2.32.0
pandas~=2.2.3
numpy~=2.1.2
opencv-python~=4.10.0.84

